object_detection_tools/requirements.txt
# basic
tqdm
Pillow
numpy
opencv-python >= 4.5.1

matplotlib

# load json
# orjson